core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
google-genai==0.2.2
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0